refactored updating of profile

+++ app/Actions/Customers/Profile/UpdateCustomerProfileAction.php
@@ -14,20 +14,26 @@ class UpdateCustomerProfileAction
     public function execute(ProfileRequest $request): JsonResponse
     {
         $data = $request->validated();
-
+        $path = null;
         /** @var Customer $customer */
         $customer = auth()->user();
-        Storage::deleteDirectory($customer->image);
+        if ($request->hasFile('profile_picture')) {
+            Storage::deleteDirectory($customer->image);
+            $path = $request->file('profile_picture')->store('customers/'.$customer->id);
+        }
 
-        $path = $request->file('profile_picture')->store('customers/'.$customer->id);
 
         $customer->update([
             'name' => $data['name'],
             'phone_number' => $data['phone_number'],
             'email' => $data['email'],
-            'image' => $path,
         ]);
 
+        if ($path) {
+            $customer->update([
+                'image' => $path,
+            ]);
+        }
         return response()->json([
             'message' => 'Profile updated successfully.',
         ]);
